integrate functions into run_pipeline
